+++ auxilium/methods/create.py
@@ -5,7 +5,7 @@
 # Python project for an automated test and deploy toolkit.
 #
 # Author:   sonntagsgesicht
-# Version:  0.2.4, copyright Wednesday, 20 October 2021
+# Version:  0.2.5, copyright Thursday, 28 October 2021
 # Website:  https://github.com/sonntagsgesicht/auxilium
 # License:  Apache License 2.0 (see LICENSE file)
 
@@ -45,7 +45,8 @@ def do(name=None, slogan=None, author=None, email=None, url=None,
         sys_path.append(project_path)
         code = code or docmaintain(pkg, path=project_path)
 
-    chdir(project_path)
+    #
+    # chdir(project_path)
     if venv:
         # clear virtual environment folder
         del_tree(venv)